classe TerrainCell passée en Polygon. Classe Cell ainsi supprimée définitivement.
La classe Polygon gère désormais aussi bien les images simples et les sprites.

+++ trunk/source/terrain/Terrain.h
@@ -28,7 +28,7 @@ namespace shootmii {
       void draw() const;
       void generate();
       bool contains(float screenX, float screenY) const;
-      int getHeight(const int screenX) const;
+      float getHeight(const int screenX) const;
       TerrainCell getGround(const int colIndex) const;
   };
